utf8parse
Struct Parser

pub struct Parser { /* private fields */ }

A parser for Utf8 Characters

Repeatedly call advance with bytes to emit Utf8 characters

Implementations

impl Parser

pub fn new() -> Parser

Create a new Parser

pub fn advance<R>(&mut self, receiver: &mut R, byte: u8)
where R: Receiver,

Advance the parser

The provider receiver will be called whenever a codepoint is completed or an invalid sequence is detected.